Cleanup MySql project
- File scope namespaces
- Primary constructor
- Reorganise files
