+++ src/TeaTime.Data.MySql/Factories/IMySqlConnectionFactory.cs
@@ -1,0 +1,8 @@
+namespace TeaTime.Data.MySql.Factories;
+
+using MySqlConnector;
+
+public interface IMySqlConnectionFactory
+{
+    MySqlConnection GetConnection();
+}
