--- src/TeaTime.Data.MySql/IMySqlConnectionFactory.cs
@@ -1,9 +1,0 @@
-namespace TeaTime.Data.MySql
-{
-    using MySqlConnector;
-
-    public interface IMySqlConnectionFactory
-    {
-        MySqlConnection GetConnection();
-    }
-}
