+++ src/TeaTime.Data.MySql/Repositories/StatisticsRepository.cs
@@ -1,26 +1,22 @@
-namespace TeaTime.Data.MySql.Repositories
+namespace TeaTime.Data.MySql.Repositories;
+
+using System.Threading.Tasks;
+using Common.Abstractions.Data;
+using Factories;
+
+public class StatisticsRepository(IMySqlConnectionFactory factory) : BaseRepository(factory), IStatisticsRepository
 {
-    using System.Threading.Tasks;
-    using Common.Abstractions.Data;
-
-    public class StatisticsRepository : BaseRepository, IStatisticsRepository
+    public Task<long> CountGlobalOrdersMadeAsync()
     {
-        public StatisticsRepository(IMySqlConnectionFactory factory) : base(factory)
-        {
-        }
+        const string sql = "SELECT count(*) FROM orders o JOIN runs r ON r.id = o.runId WHERE r.ended = 1";
 
-        public Task<long> CountGlobalOrdersMadeAsync()
-        {
-            const string sql = "SELECT count(*) FROM orders o JOIN runs r ON r.id = o.runId WHERE r.ended = 1";
+        return SingleOrDefaultAsync<long>(sql);
+    }
 
-            return SingleOrDefaultAsync<long>(sql);
-        }
+    public Task<long> CountGlobalEndedRunsAsync()
+    {
+        const string sql = "SELECT count(*) FROM runs WHERE ended = 1";
 
-        public Task<long> CountGlobalEndedRunsAsync()
-        {
-            const string sql = "SELECT count(*) FROM runs WHERE ended = 1";
-
-            return SingleOrDefaultAsync<long>(sql);
-        }
+        return SingleOrDefaultAsync<long>(sql);
     }
 }
